# studio/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LazyEdit private Studio API",
    "version": "1.0.0",
    "description": "One configured owner pilot. See handoff for device-code errors and limitations. No implicit publication."
  },
  "servers": [
    {
      "url": "https://edit.lazying.art"
    }
  ],
  "security": [
    {
      "bearer": []
    }
  ],
  "paths": {
...
    "/auth/login": {
      "post": {
        "summary": "Password exchange; mode token returns a scoped token pair",
        "responses": {
          "200": {
            "description": "Success (completion is distinct from job acceptance)"
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Missing, expired or revoked user credential"
          },
          "403": {
            "description": "Scope, ownership or browser-origin denied"
          },
          "409": {
            "description": "Offset, idempotency or reviewed artifact conflict"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "security": []
      }
    },
    "/auth/device": {
      "post": {
        "summary": "Create device authorization request",
        "responses": {
          "200": {
            "description": "Success (completion is distinct from job acceptance)"
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Missing, expired or revoked user credential"
          },
          "403": {
            "description": "Scope, ownership or browser-origin denied"
          },
          "409": {
            "description": "Offset, idempotency or reviewed artifact conflict"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "security": []
      }
    },
    "/auth/token": {
      "post": {
        "summary": "Refresh token rotation or poll approved device code",
        "responses": {
          "200": {
            "description": "Success (completion is distinct from job acceptance)"
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Missing, expired or revoked user credential"
          },
          "403": {
            "description": "Scope, ownership or browser-origin denied"
          },
          "409": {
            "description": "Offset, idempotency or reviewed artifact conflict"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "security": []
      }
    },
...
    "/auth/logout": {
      "post": {
        "summary": "logout",
        "responses": {
          "200": {
            "description": "Success (completion is distinct from job acceptance)"
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Missing, expired or revoked user credential"
          },
          "403": {
            "description": "Scope, ownership or browser-origin denied"
          },
          "409": {
            "description": "Offset, idempotency or reviewed artifact conflict"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Opaque short-lived user token; not a LazyEdge machine token"
      }
    }
  }
}